+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - PR #3161 Move merge files to legacy
 - PR #3079 Added support to write ORC files given a local path
 - PR #3192 Add dtype param to cast `DataFrame` on init
+- PR #3222 Add nvtext character tokenizer
 
 ## Improvements